Ajout tests unitaires v3

+++ TP1/prog1.py
@@ -7,8 +7,8 @@ def carre():
     a = int(input("Put a number: "))
     print(a**2)
 
-a = float(input("Nombre: "))
-b = float(input("Puissance: "))
+a = int(input("Nombre: "))
+b = int(input("Puissance: "))
 
 res = f.puissance(a, b)
 print(f"Result: {res}")
